+++ pkgs/dart_mcp/CHANGELOG.md
@@ -1,9 +1,16 @@
 ## 0.2.0-wip
 
-- **Breaking**: Fixed paginated result subtypes to use `nextCursor` instead of
-  `cursor` as the key for the next cursor.
+- Support protocol version 2025-03-26.
+  - Adds support for `AudioContent`.
+  - Adds support for `ToolAnnotations`.
+  - Adds support for `ProgressNotification` messages.
 - Save the `ServerCapabilities` object on the `ServerConnection` class to make
   it easier to check the capabilities of the server.
+- **Breaking**: Fixed paginated result subtypes to use `nextCursor` instead of
+  `cursor` as the key for the next cursor.
+- **Breaking**: Change the `ProgressNotification.progress` and
+  `ProgressNotification.total` types to `num` instead of `int` to align with the
+  spec.
 
 ## 0.1.0